Update Project 7
I had to write a function to reverse the element of an integer 1-D array and return the reversed array. Display the original array and reversed array on the screen.

For example, the input array is {1,3,5,7,9}, output would be {9,7,5,3,1}.

Write a function to find the smallest element of an array and return this element.

For example, the input is {1,5,9,4,0,3,3,7}, output would be 0.